Migrate localization system to JSON and update remaining strings
- Implemented `i18n.py` with `SimpleTranslator` class to load translations from JSON files.
- Replaced `.po`/`.mo` files with `locale/en.json` and `locale/fr.json`.
- Updated `japanese_examples.py` to use `i18n.py` for error messages and status updates, replacing hardcoded strings with localized keys (`error_tatoeba_connection`, `no_japanese_sentence_found`).
- Updated `GUI.py` to remove `gettext` dependency and use `i18n.py`.
- Updated `tests/test_gui_async.py` to remove obsolete mocks and fix import paths.
