+++ locale/en.json
@@ -7,5 +7,7 @@
     "no_example_selected": "No example selected.",
     "{DST_FIELD_JAP}_field_not_found": "{DST_FIELD_JAP} field not found in the current note.",
     "{DST_FIELD_TRANSLATION}_field_not_found": "{DST_FIELD_TRANSLATION} field not found in the current note.",
-    "add_example_manually_tip": "Add an example sentence manually"
+    "add_example_manually_tip": "Add an example sentence manually",
+    "error_tatoeba_connection": "Error: Unable to connect to Tatoeba API.",
+    "no_japanese_sentence_found": "No Japanese sentence found containing the word '{word}'."
 }
