+++ locale/fr.json
@@ -7,5 +7,7 @@
     "no_example_selected": "Aucun exemple sélectionné.",
     "{DST_FIELD_JAP}_field_not_found": "Champs {DST_FIELD_JAP} non trouvé dans la note en cours d'édition.",
     "{DST_FIELD_TRANSLATION}_field_not_found": "Champs {DST_FIELD_TRANSLATION} non trouvé dans la note en cours d'édition.",
-    "add_example_manually_tip": "Ajouter un exemple manuellement"
+    "add_example_manually_tip": "Ajouter un exemple manuellement",
+    "error_tatoeba_connection": "Erreur : Impossible de se connecter à l'API Tatoeba.",
+    "no_japanese_sentence_found": "Aucune phrase japonaise trouvée contenant le mot '{word}'."
 }
